Define method to place ships on grid
Class Grid has now a method "distribute_ships" to place the right number
of every ship type on the grid at a random place. Therefore, it uses a helper function
"_place_ship" that checks, if the ship fits to that random position and
places it in case it does.

PROBLEM: Ships can be placed next to each other. How to distinguish them
later on?...