Define Grid.shoot method
A computer or user move is now processed correctly. The method checks,
if the field was already guessed and if this is not the case, it looks
for the entry of the field. In case of a ship, the method checks if the
ship was destroyed. If the ship is destroyed, the method checks if all
ships are destroyed and sets the game_over attribute to True.

Two helper functions are introduced:
_get_ship_size  returns an int with the size of the ship that occupies
the field
_is_destroyed   returns a bool stating if the ship occupying the field
is destroyed